hd-config-file: fix segfault
The second file passed in the callback can be NULL, and is only used
when the file is renamed. I think this bug was introduced by the GIO
port done in 3391b3a4c7a3b504d46ab695855b9d77d3e892b7
("Removed GNOMEVFS dependency (converted to GIO)")

Also clarify the names of the file arguments while at it.